+++ src/socketlib/basic/receive.py
@@ -1,39 +1,29 @@
 import logging
 import queue
+import socket
 from typing import Callable, Optional
 
 from socketlib.basic.buffer import Buffer
 from socketlib.basic.queues import put_in_queue
 
 
-def get_msg(buffer: Buffer, msg_end: bytes) -> Optional[bytes]:
+def get_msg(
+        buffer: Buffer,
+        msg_end: bytes,
+        logger: Optional[logging.Logger] = None,
+        name: str = ""
+) -> Optional[bytes]:
     """ Get a message from a socket buffer.
     """
     try:
         return buffer.get_msg(msg_end=msg_end)
     except ConnectionError:
-        return
+        error = f"{name} failed to get message. Connection lost"
+    except socket.timeout:
+        error = f"{name} failed to get message. Timed out"
 
-
-def receive_msg(
-        buffer: Buffer,
-        msg_queue: queue.Queue,
-        msg_end: bytes,
-        logger: Optional[logging.Logger] = None,
-        name: str = ""
-) -> bool:
-    """ Receive a message from a socket.
-
-        Returns True if there is an error.
-    """
-    data = get_msg(buffer, msg_end)
-    if data is not None:
-        msg_queue.put(data)
-    else:
-        if logger:
-            logger.info(f"{name} failed to receive message")
-        return True
-    return False
+    if error and logger:
+        logger.info(error)
 
 
 def receive_and_enqueue(
@@ -48,12 +38,10 @@ def receive_and_enqueue(
     """ Receive a message and put it in a queue
     """
     while not stop():
-        msg = get_msg(buffer, msg_end)
+        msg = get_msg(buffer, msg_end, logger, name)
         if msg is not None:
             success = put_in_queue(msg, msg_queue, timeout)
             if not success and logger:
                 logger.info(f"{name} failed to enqueue message")
         else:
-            if logger:
-                logger.info(f"{name} failed to receive message")
             break